@@ -3,19 +3,19 @@ version: 4.1.0
 description: A CMIS 1.1 compliant browser and server based client for Dart.
 homepage: https://github.com/shamblett/cmis
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dependencies:
-  json_object_lite: '^2.0.0'
-  http: '^0.12.0+2'
-  mime: '^0.9.6+3'
+  json_object_lite: '^2.1.1'
+  http: '^0.12.2'
+  mime: '^1.0.0-nullsafety.0'
 
 dev_dependencies:
   bootjack: '^1.0.1'
-  test: '^1.14.2'
-  pana: '^0.13.5'
-  pedantic: '^1.9.0'
-  build_runner: '^1.8.0'
-  build_test: '^1.2.0'
-  build_web_compilers: '^2.8.0'
+  test: '^1.16.0-nullsafety.16'
+  pana: '^0.14.10'
+  pedantic: '^1.10.0-nullsafety.3'
+  build_runner: '^1.10.13'
+  build_test: '^1.3.4'
+  build_web_compilers: '^2.16.0'
 
