@@ -12,22 +12,22 @@ import 'package:json_object_lite/json_object_lite.dart';
 import 'package:cmis/cmis_server_client.dart';
 import 'package:cmis/cmis.dart';
 
-CmisSession cmisSession;
+CmisSession? cmisSession;
 
 // Cmis get repositories function wrapped in a future wrapper
 Future<JsonObjectLite<dynamic>> getRepositories() async {
   final completer = Completer<JsonObjectLite<dynamic>>();
 
   void localcompleter() {
-    final dynamic cmisResponse = cmisSession.completionResponse;
+    final dynamic cmisResponse = cmisSession!.completionResponse;
     final dynamic repo = cmisResponse.jsonCmisResponse.toList()[0];
-    final String repositoryId = repo.toList()[0];
-    cmisSession.repositoryId = repositoryId;
+    final String? repositoryId = repo.toList()[0];
+    cmisSession!.repositoryId = repositoryId;
     completer.complete(repo);
   }
 
-  cmisSession.resultCompletion = localcompleter;
-  cmisSession.getRepositories();
+  cmisSession!.resultCompletion = localcompleter;
+  cmisSession!.getRepositories();
   return completer.future;
 }
 
@@ -55,5 +55,5 @@ Future<void> main() async {
   // Get the repositories, future based
   await getRepositories();
   // or ... JsonObjectLite<dynamic> info = await getRepositories();
-  print('Repository id is : ${cmisSession.repositoryId}');
+  print('Repository id is : ${cmisSession!.repositoryId}');
 }
