@@ -13,12 +13,12 @@
 
 part of cmisbrowsertest;
 
-TextAreaElement queryStatement = querySelector('#cmis-query-statement');
-DivElement queryAlertSection = querySelector('#cmis-alertsection-query');
-DivElement queryListSection = querySelector('#cmis-query-list-section');
+TextAreaElement? queryStatement = querySelector('#cmis-query-statement') as TextAreaElement?;
+DivElement? queryAlertSection = querySelector('#cmis-alertsection-query') as DivElement?;
+DivElement? queryListSection = querySelector('#cmis-query-list-section') as DivElement?;
 void doQueryClear(Event e) {
-  queryListSection.children.clear();
-  queryAlertSection.children.clear();
+  queryListSection!.children.clear();
+  queryAlertSection!.children.clear();
 }
 
 void outputQueryList(dynamic response) {
@@ -29,30 +29,30 @@ void outputQueryList(dynamic response) {
     final results = response.jsonCmisResponse.results;
     if (results.isNotEmpty) {
       preElement.innerHtml = results.toString();
-      queryListSection.children.add(preElement);
+      queryListSection!.children.add(preElement);
     } else {
       final noChildren = LIElement();
       noChildren.innerHtml = 'The query has returned no items';
       uList.children.add(noChildren);
-      queryListSection.children.add(uList);
+      queryListSection!.children.add(uList);
     }
   } else {
     final noChildren = LIElement();
     noChildren.innerHtml = 'There query has returned no items';
     uList.children.add(noChildren);
-    queryListSection.children.add(uList);
+    queryListSection!.children.add(uList);
   }
 }
 
 void doQuery(Event e) {
   void completer() {
-    final dynamic cmisResponse = cmisSession.completionResponse;
+    final dynamic cmisResponse = cmisSession!.completionResponse;
 
     if (cmisResponse.error) {
       final dynamic errorResponse = cmisResponse.jsonCmisResponse;
-      final int errorCode = cmisResponse.errorCode;
-      String error;
-      String reason;
+      final int? errorCode = cmisResponse.errorCode;
+      String? error;
+      String? reason;
       if (errorCode == 0) {
         error = errorResponse.error;
         reason = errorResponse.reason;
@@ -62,17 +62,17 @@ void doQuery(Event e) {
       }
 
       final message = 'Error - $error, Reason - $reason, Code - $errorCode';
-      addErrorAlert(queryAlertSection, message);
+      addErrorAlert(queryAlertSection!, message);
     } else {
       outputQueryList(cmisResponse);
     }
   }
 
-  clearAlertSection(queryAlertSection);
-  cmisSession.resultCompletion = completer;
-  if (queryStatement.value.isEmpty) {
-    addErrorAlert(queryAlertSection, 'You must supply a query!');
+  clearAlertSection(queryAlertSection!);
+  cmisSession!.resultCompletion = completer;
+  if (queryStatement!.value!.isEmpty) {
+    addErrorAlert(queryAlertSection!, 'You must supply a query!');
   } else {
-    cmisSession.query(queryStatement.value.trim());
+    cmisSession!.query(queryStatement!.value!.trim());
   }
 }
